requirements.txt
asgiref
certifi
cffi
charset-normalizer
crispy-bootstrap5
cryptography
defusedxml
distlib
dj-database-url
dj-rest-auth
Django
django-allauth
django-cors-headers
django-crispy-forms
django-email
django-environ
django-filter
djangorestframework
filelock
idna
Markdown
MouseInfo
numpy
oauthlib
opencv-python
pillow
pipenv
platformdirs
psycopg2
psycopg2-binary
PyAutoGUI
pycparser
PyGetWindow
PyJWT
PyMsgBox
pyperclip
PyRect
PyScreeze
python3-openid
pytweening
pytz
requests
requests-oauthlib
setuptools
sqlparse
typing_extensions
tzdata
urllib3
virtualenv